bug fixes in GNSS functions
